Creating algoritm to draw a robot symbol on canvas

+++ src/sample/Detector.java
@@ -6,7 +6,7 @@ package sample;
 public class Detector {
     public Detector(){}
 
-    public static Signal measureSignal(int robot_x, int robot_y, Antena ant){
+    public static Signal measureSignal(double robot_x, double robot_y, Antena ant){
         Signal result = new Signal();
         result.setAntID(ant.getID());
         int ant_x = ant.getX();
